+++ src/main/java/matteroverdrive/registry/ModCreativeTabs.java
@@ -22,7 +22,6 @@ public final class ModCreativeTabs {
                         OverhaulContent.BLOCK_ITEMS.values().forEach(item -> output.accept(item.get()));
                         output.accept(ModItems.DRONE_DEPLOYMENT_CORE.get());
                         output.accept(ModExoticItems.VEX_MYTHOCLAST.get());
-                        ModDestinyItems.WEAPONS.values().forEach(item -> output.accept(item.get()));
                     })
                     .build());